requirements.txt
streamlit
pandas
numpy
ta
smartapi-python
pyotp
requests